+++ src/Tabby/Core/Input/Input.h
@@ -21,8 +21,14 @@ class Input {
 public:
     static void Init();
 
-    static bool IsKeyPressed(KeyCode key);
-    static bool IsMouseButtonPressed(MouseCode button);
+    static bool GetKeyDown(KeyCode key);
+    static bool GetKeyUp(KeyCode key);
+    static bool GetKey(KeyCode key);
+    static bool GetMouseButtonDown(MouseCode button);
+    static bool GetMouseButtonUp(MouseCode button);
+    static bool GetMouseButton(MouseCode button);
+    // static bool IsKeyPressed(KeyCode key);
+    // static bool IsMouseButtonPressed(MouseCode button);
     static Vector2 GetMousePosition();
     static Vector2 GetMouseScrollDelta();
     static float GetMouseX();
@@ -37,9 +43,14 @@ public:
 
 private:
     Input();
+    static void Update();
 
     GamepadInfo* m_Gamepads;
     Vector2 m_MouseScrollDelta;
+    std::unordered_map<KeyCode, bool> currentKeyStates;
+    std::unordered_map<KeyCode, bool> previousKeyStates;
+    std::unordered_map<MouseCode, bool> currentMouseButtonStates;
+    std::unordered_map<MouseCode, bool> previousMouseButtonStates;
 
 private:
     inline static Input* s_Instance;