+++ src/Tabby/World/Components.h
@@ -63,6 +63,37 @@ struct TransformComponent {
     operator Matrix4&() { return TransformMatrix; }
 };
 
+// struct TransformComponent {
+// public:
+//     Vector3 position = { 0.0f, 0.0f, 0.0f };
+//     Vector3 rotation = { 0.0f, 0.0f, 0.0f };
+//     Vector3 scale = { 1.0f, 1.0f, 1.0f };
+//
+//     Matrix4 worldTransformMatrix = Matrix4(1);
+//     Matrix4 transformMatrix = Matrix4(1);
+//
+//     TransformComponent() = default;
+//     TransformComponent(const TransformComponent&) = default;
+//     TransformComponent(const Vector3& Position)
+//         : position(Position)
+//     {
+//     }
+//
+//     Matrix4& GetTransform() { return transformMatrix; }
+//     const Matrix4& GetTransform() const { return transformMatrix; }
+//     Matrix4& GetWorldTransform() { return worldTransformMatrix; }
+//     const Matrix4& GetWorldTransform() const { return worldTransformMatrix; }
+//     void ApplyTransform(const Matrix4& transform);
+//     // void ApplyTransformToLocal(const Matrix4& transform);
+//
+//     operator Matrix4&() { return worldTransformMatrix; }
+//
+// private:
+//     Vector3 worldPosition = { 0.0f, 0.0f, 0.0f };
+//     Vector3 worldRotation = { 0.0f, 0.0f, 0.0f };
+//     Vector3 worldScale = { 1.0f, 1.0f, 1.0f };
+// };
+
 struct SpriteRendererComponent {
     Vector4 Color { 1.0f, 1.0f, 1.0f, 1.0f };
     AssetHandle Texture;